Update pagination to accept cocktails as children
