+++ src/pages/Cocktails.tsx
@@ -1,9 +1,8 @@
-import { Link } from 'react-router-dom'
 import LoadingGif from '../assets/loading.gif'
 import { CocktailData } from '../types/CocktailData'
 import styled from 'styled-components'
 import { Helmet } from 'react-helmet'
-import { useEffect, useState } from 'react'
+import { useState } from 'react'
 import Pagination from '../components/Pagination'
 
 interface CocktailProps {
@@ -13,54 +12,32 @@ interface CocktailProps {
 
 export default function Cocktails({ cocktailData, loading }: CocktailProps) {
     const [searchTerm, setSearchTerm] = useState('');
-    const [currentPage, setCurrentPage] = useState(1);
-    const itemsPerPage = 9;
-
-    useEffect(() => {
-        setCurrentPage(1);
-    }, [searchTerm]);
-
+    
     const filteredCocktails = cocktailData.filter(cocktail =>
         cocktail.name.toLowerCase().includes(searchTerm.toLowerCase())
     );
 
-    const indexOfLastItem = currentPage * itemsPerPage;
-    const indexOfFirstItem = indexOfLastItem - itemsPerPage;
-    const currentCocktails = filteredCocktails.slice(indexOfFirstItem, indexOfLastItem);
-
-    const renderCocktails = currentCocktails.map(cocktail => {
-        const {id, name, glass, slug, image_url} = cocktail;
-        return (
-            <CocktailCard key={id}>
-                <Link to={`/cocktail/${slug}`} state={{ from: "cocktail", data: cocktail}}>
-                    <h3>{name}</h3>
-                    <img src={image_url} alt={`${name} inside a ${glass} glass`} />
-                </Link>
-            </CocktailCard>
-        );
-    });
-
     return (
         <>
             <Wrapper>
-                <h2>Cocktails</h2>
-                <Input
-                    type="text"
-                    placeholder="Search cocktails..."
-                    value={searchTerm}
-                    onChange={(e) => setSearchTerm(e.target.value)}
-                />
+                <Header>
+                    <h2>Cocktails</h2>
+                    <Input
+                        type="text"
+                        placeholder="Search cocktails..."
+                        value={searchTerm}
+                        onChange={(e) => setSearchTerm(e.target.value)}
+                    />
+                    <button>Toggle Fridge</button>
+                    <button>Toggle Recommendations</button>
+                </Header> 
                 {loading ? <img src={LoadingGif} className="loader" alt="loader" /> : (
                     <>
-                        <CocktailListContainer>
-                            {renderCocktails}
-                        </CocktailListContainer>
+                        
                         <Pagination
                             totalItems={filteredCocktails.length}
-                            itemsPerPage={itemsPerPage}
-                            currentPage={currentPage}
-                            onPageChange={setCurrentPage}
-                        />
+                            searchTerm={searchTerm}
+                        >{filteredCocktails}</Pagination>
                     </>
                 )}
             </Wrapper>
@@ -83,49 +60,10 @@ const Wrapper = styled.div`
     align-self: flex-start;
 `
 
-const CocktailListContainer = styled.div`
-
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: 15px;
-
-    @media only screen and (min-width:1025px) {
-      flex-direction: row;
-      flex-wrap: wrap;
-    }
-`
 
 
-const CocktailCard = styled.a`
-    a {
-        width: 80vw;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        background-color: #838383;
-        border-radius: 1vh;
-        margin: 5px;
-        
-    }
-  
-    a > img{
-        width: 100%;
-        border-radius: 1vh;
-    }
 
-    @media only screen and (min-width:1025px) {
-        a{
-          width: 25vw;
-        }
-        
-        a >img{
-          width: 25vw;
-          height: 18vw;
-        }
-      }
-`
+
 
 const Input = styled.input`
   margin: 20px;
@@ -141,3 +79,11 @@ const Input = styled.input`
     box-shadow: 0 0 5px rgba(0,123,255,0.5);
   }
 `;
+
+const Header = styled.div`
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 10px;
+    margin: 20px;
+    `
